requirements.txt
spotipy
streamlit
plotly 
python-dotenv